refactor includes and tidy build scripts
