+++ include/boost/buffers/mutable_buffer_span.hpp
@@ -11,8 +11,7 @@
 #define BOOST_BUFFERS_MUTABLE_BUFFER_SPAN_HPP
 
 #include <boost/buffers/detail/config.hpp>
-#include <boost/buffers/mutable_buffer.hpp>
-#include <boost/buffers/tag_invoke.hpp>
+#include <boost/buffers/buffer.hpp>
 #include <boost/buffers/type_traits.hpp>
 
 namespace boost {
