+++ include/boost/buffers/range.hpp
@@ -11,137 +11,12 @@
 #define BOOST_BUFFERS_RANGE_HPP
 
 #include <boost/buffers/detail/config.hpp>
-#include <boost/buffers/const_buffer.hpp>
-#include <boost/buffers/mutable_buffer.hpp>
+#include <boost/buffers/buffer.hpp>
 #include <boost/buffers/type_traits.hpp>
 #include <type_traits>
 
 namespace boost {
 namespace buffers {
-
-/** Return an iterator to the beginning of the buffer sequence.
-*/
-constexpr struct
-{
-    // mutable_buffer is a sequence of length 1
-    template<class T>
-    auto operator()(T const& t) const noexcept -> typename std::enable_if<
-        std::is_convertible<T const*, mutable_buffer const*>::value,
-        mutable_buffer const*>::type
-    {
-        return static_cast<mutable_buffer const*>(std::addressof(t));
-    }
-
-    // const_buffer is a sequence of length 1
-    template<class T>
-    auto operator()(T const& t) const noexcept -> typename std::enable_if<
-        std::is_convertible<T const*, const_buffer const*>::value,
-        const_buffer const*>::type
-    {
-        return static_cast<const_buffer const*>(std::addressof(t));
-    }
-
-    // Convertible T is a sequence of length 1
-    template<class T>
-    auto operator()(T const& t) const noexcept ->
-        typename std::enable_if<
-            ! std::is_convertible<T const*, const_buffer const*>::value &&
-            ! std::is_convertible<T const*, mutable_buffer const*>::value && (
-                std::is_convertible<T, const_buffer>::value ||
-                std::is_convertible<T, mutable_buffer>::value)
-        >::type
-    {
-        return std::addressof(t);
-    }
-
-    // Bidirectional range with convertible value type
-    template<class T>
-    auto operator()(T& t) const noexcept -> typename std::enable_if<
-        ! std::is_convertible<T const*, const_buffer const*>::value &&
-        ! std::is_convertible<T const*, mutable_buffer const*>::value &&
-        ! std::is_convertible<T, const_buffer>::value &&
-        ! std::is_convertible<T, mutable_buffer>::value &&
-        detail::is_bidirectional_iterator<decltype(t.begin())>::value,
-        decltype(t.begin())>::type
-    {
-        return t.begin();
-    }
-
-    // Bidirectional range with convertible value type
-    template<class T>
-    auto operator()(T const& t) const noexcept -> typename std::enable_if<
-        ! std::is_convertible<T const*, const_buffer const*>::value &&
-        ! std::is_convertible<T const*, mutable_buffer const*>::value &&
-        ! std::is_convertible<T, const_buffer>::value &&
-        ! std::is_convertible<T, mutable_buffer>::value &&
-        detail::is_bidirectional_iterator<decltype(t.begin())>::value,
-        decltype(t.begin())>::type
-    {
-        return t.begin();
-    }
-} begin {};
-
-//------------------------------------------------------------------------------
-
-/** Return an iterator to the end of the buffer sequence.
-*/
-constexpr struct
-{
-    // mutable_buffer is a sequence of length 1
-    template<class T>
-    auto operator()(T const& t) const noexcept -> typename std::enable_if<
-        std::is_convertible<T const*, mutable_buffer const*>::value,
-        mutable_buffer const*>::type
-    {
-        return static_cast<mutable_buffer const*>(std::addressof(t)) + 1;
-    }
-
-    // const_buffer is a sequence of length 1
-    template<class T>
-    auto operator()(T const& t) const noexcept -> typename std::enable_if<
-        std::is_convertible<T const*, const_buffer const*>::value,
-        const_buffer const*>::type
-    {
-        return static_cast<const_buffer const*>(std::addressof(t)) + 1;
-    }
-
-    // Convertible T is a sequence of length 1
-    template<class T>
-    auto operator()(T const& t) const noexcept ->
-        typename std::enable_if<
-            ! std::is_convertible<T const*, const_buffer const*>::value &&
-            ! std::is_convertible<T const*, mutable_buffer const*>::value && (
-                std::is_convertible<T, const_buffer>::value ||
-                std::is_convertible<T, mutable_buffer>::value)
-        >::type
-    {
-        return std::addressof(t) + 1;
-    }
-
-    // Bidirectional range with convertible value type
-    template<class T>
-    auto operator()(T& t) const noexcept -> typename std::enable_if<
-        ! std::is_convertible<T const*, const_buffer const*>::value &&
-        ! std::is_convertible<T const*, mutable_buffer const*>::value &&
-        ! std::is_convertible<T, const_buffer>::value &&
-        ! std::is_convertible<T, mutable_buffer>::value,
-        decltype(t.end())>::type
-    {
-        return t.end();
-    }
-
-    // Bidirectional range with convertible value type
-    template<class T>
-    auto operator()(T const& t) const noexcept -> typename std::enable_if<
-        ! std::is_convertible<T const*, const_buffer const*>::value &&
-        ! std::is_convertible<T const*, mutable_buffer const*>::value &&
-        ! std::is_convertible<T, const_buffer>::value &&
-        ! std::is_convertible<T, mutable_buffer>::value,
-        decltype(t.end())>::type
-    {
-        return t.end();
-    }
-} end {};
 
 //------------------------------------------------
