+++ include/boost/buffers/string_buffer.hpp
@@ -10,8 +10,7 @@
 #ifndef BOOST_BUFFERS_STRING_BUFFER_HPP
 #define BOOST_BUFFERS_STRING_BUFFER_HPP
 
-#include <boost/buffers/const_buffer.hpp>
-#include <boost/buffers/mutable_buffer.hpp>
+#include <boost/buffers/buffer.hpp>
 #include <boost/buffers/detail/except.hpp>
 #include <boost/assert.hpp>
 #include <string>
