+++ include/boost/buffers/type_traits.hpp
@@ -11,163 +11,10 @@
 #define BOOST_BUFFERS_TYPE_TRAITS_HPP
 
 #include <boost/buffers/detail/config.hpp>
-#include <boost/buffers/detail/type_traits.hpp>
-#include <type_traits>
+#include <boost/buffers/buffer.hpp>
 
 namespace boost {
 namespace buffers {
-
-class const_buffer;
-class mutable_buffer;
-
-// https://www.boost.org/doc/libs/1_65_0/doc/html/boost_asio/reference/ConstBufferSequence.html
-
-/** Determine if T is a ConstBuffers.
-*/
-template<class T, class = void>
-struct is_const_buffer_sequence
-    : std::false_type
-{
-};
-
-template<class T>
-struct is_const_buffer_sequence<T const>
-    : is_const_buffer_sequence<typename
-        std::decay<T>::type>
-{
-};
-
-template<class T>
-struct is_const_buffer_sequence<T const&>
-    : is_const_buffer_sequence<typename
-        std::decay<T>::type>
-{
-};
-
-template<class T>
-struct is_const_buffer_sequence<T&>
-    : is_const_buffer_sequence<typename
-        std::decay<T>::type>
-{
-};
-
-template<>
-struct is_const_buffer_sequence<
-    const_buffer>
-    : std::true_type
-{
-};
-
-template<>
-struct is_const_buffer_sequence<
-    mutable_buffer>
-    : std::true_type
-{
-};
-
-template<class T>
-struct is_const_buffer_sequence<T, void_t<
-    typename std::enable_if<
-        (std::is_same<const_buffer, typename 
-            T::value_type>::value
-        || std::is_same<mutable_buffer, typename
-            T::value_type>::value
-            ) &&
-        detail::is_bidirectional_iterator<typename
-            T::const_iterator>::value &&
-        std::is_same<typename
-            T::const_iterator, decltype(
-            std::declval<T const&>().begin())
-                >::value &&
-        std::is_same<typename
-            T::const_iterator, decltype(
-            std::declval<T const&>().end())
-                >::value && (
-        std::is_same<const_buffer, typename
-            std::remove_const<typename
-                std::iterator_traits<
-                    typename T::const_iterator
-                        >::value_type>::type
-                >::value ||
-        std::is_same<mutable_buffer, typename
-            std::remove_const<typename
-                std::iterator_traits<
-                    typename T::const_iterator
-                        >::value_type>::type
-                >::value)
-        // VFALCO This causes problems when the
-        // trait is used to constrain ctors
-        // && std::is_move_constructible<T>::value
-            >::type
-    > > : std::true_type
-{
-};
-
-/** Determine if T is a MutableBuffers.
-*/
-template<class T, class = void>
-struct is_mutable_buffer_sequence : std::false_type
-{
-};
-
-template<class T>
-struct is_mutable_buffer_sequence<T const>
-    : is_mutable_buffer_sequence<typename
-        std::decay<T>::type>
-{
-};
-
-template<class T>
-struct is_mutable_buffer_sequence<T const&>
-    : is_mutable_buffer_sequence<typename
-        std::decay<T>::type>
-{
-};
-
-template<class T>
-struct is_mutable_buffer_sequence<T&>
-    : is_mutable_buffer_sequence<typename
-        std::decay<T>::type>
-{
-};
-
-template<>
-struct is_mutable_buffer_sequence<
-    mutable_buffer>
-    : std::true_type
-{
-};
-
-template<class T>
-struct is_mutable_buffer_sequence<T, void_t<
-    typename std::enable_if<
-        std::is_same<mutable_buffer, typename
-            T::value_type>::value &&
-        detail::is_bidirectional_iterator<typename
-            T::const_iterator>::value &&
-        std::is_same<typename
-            T::const_iterator, decltype(
-            std::declval<T const&>().begin())
-                >::value &&
-        std::is_same<typename
-            T::const_iterator, decltype(
-            std::declval<T const&>().end())
-                >::value &&
-        std::is_same<mutable_buffer, typename
-            std::remove_const<typename
-                std::iterator_traits<
-                    typename T::const_iterator
-                        >::value_type>::type
-                >::value
-        // VFALCO This causes problems when the
-        // trait is used to constrain ctors
-        // && std::is_move_constructible<T>::value
-            >::type
-    >> : std::true_type
-{
-};
-
-//------------------------------------------------
 
 /** Determine if T is a DynamicBuffer
 */
